+++ quality.sh
@@ -47,7 +47,7 @@ echo "🏗️ Building release version..."
 cargo build --release
 
 echo "🔍 Running DenoJS tests..."
-deno test --allow-read tests/*.ts
+deno test --allow-read --allow-env tests/*.ts
 
 echo "📝 Formatting JS, HTML, and CSS files with deno fmt..."
 deno fmt docs/*.js docs/*.html docs/*.css helpers/*.ts tests/*.ts